Move CompleteInferredSpan to unexported method on LifecycleProcessor.
