+++ pkg/serverless/trace/inferredspan/inferred_span.go
@@ -18,9 +18,6 @@ import (
 	pb "github.com/DataDog/datadog-agent/pkg/proto/pbgo/trace"
 	"github.com/DataDog/datadog-agent/pkg/serverless/random"
 	"github.com/DataDog/datadog-agent/pkg/serverless/tags"
-	"github.com/DataDog/datadog-agent/pkg/trace/api"
-	"github.com/DataDog/datadog-agent/pkg/trace/info"
-	"github.com/DataDog/datadog-agent/pkg/trace/sampler"
 	"github.com/DataDog/datadog-agent/pkg/util/log"
 )
 
@@ -98,46 +95,6 @@ func FilterFunctionTags(input map[string]string) map[string]string {
 	return output
 }
 
-// CompleteInferredSpan finishes the inferred span and passes it
-// as an API payload to be processed by the trace agent
-func (inferredSpan *InferredSpan) CompleteInferredSpan(
-	processTrace func(p *api.Payload),
-	endTime time.Time,
-	isError bool,
-	traceID uint64,
-	samplingPriority sampler.SamplingPriority) {
-
-	durationIsSet := inferredSpan.Span.Duration != 0
-	if inferredSpan.IsAsync {
-		// SNSSQS span duration is set in invocationlifecycle/init.go
-		if !durationIsSet {
-			inferredSpan.Span.Duration = inferredSpan.CurrentInvocationStartTime.UnixNano() - inferredSpan.Span.Start
-		}
-	} else {
-		inferredSpan.Span.Duration = endTime.UnixNano() - inferredSpan.Span.Start
-	}
-	if isError {
-		inferredSpan.Span.Error = 1
-	}
-
-	inferredSpan.Span.TraceID = traceID
-
-	traceChunk := &pb.TraceChunk{
-		Origin:   "lambda",
-		Spans:    []*pb.Span{inferredSpan.Span},
-		Priority: int32(samplingPriority),
-	}
-
-	tracerPayload := &pb.TracerPayload{
-		Chunks: []*pb.TraceChunk{traceChunk},
-	}
-
-	processTrace(&api.Payload{
-		Source:        info.NewReceiverStats().GetTagStats(info.Tags{}),
-		TracerPayload: tracerPayload,
-	})
-}
-
 // GenerateSpanId creates a secure random span id in specific scenarios, otherwise return a pseudo random id
 func GenerateSpanId() uint64 {
 	isSnapStart := os.Getenv(tags.InitType) == tags.SnapStartValue
@@ -152,9 +109,9 @@ func GenerateSpanId() uint64 {
 	return random.Random.Uint64()
 }
 
-// generateInferredSpan declares and initializes a new inferred span
-// with the SpanID and TraceID
-func (inferredSpan *InferredSpan) generateInferredSpan(startTime time.Time) {
+// GenerateInferredSpan declares and initializes a new inferred span with a
+// SpanID
+func (inferredSpan *InferredSpan) GenerateInferredSpan(startTime time.Time) {
 
 	inferredSpan.CurrentInvocationStartTime = startTime
 	inferredSpan.Span = &pb.Span{
